requirements.txt
torch
torchvision
torchaudio
matplotlib
numpy
pandas
tqdm
scikit-learn
jupyter
pillow
kagglehub
streamlit
